refactor: extract BudgetBoard component; update store, test and interface
