+++ src/interfaces/cartItem.d.ts
@@ -2,6 +2,7 @@ import { IDesign } from "@/interfaces/design";
 import { PriceBar } from "@/store/reducer/priceBarSlice";
 
 interface ICartItem {
+  user_id: string;
   design: IDesign;
   quotation: string;
   quotationDetails: PriceBar[];
